🛠 day-14 cleanup
Switched from [[Int]] to Map (Int,Int) Int but that isn't the bottleneck
that makes this whole thing slow. The culprit is likely `knothash`.

Apparently the graph building and algorithms take negligible time.

If main prints out the hashes and does nothing else the program takes
about the same amount of time to execute.