🛠 day-16 another failed attempt at part 2
The `Partner` moves make the permutation, in principle, change at each
of the iterations of a whole dance!

This calculation computes the first round of dance as a permutation.
Then mods 1.000.000.000 by its period, in hope of finding the small
remainder that is the number of times to replicate the permutation.

ALAS as I said, the approach is flawed.

Need to search for the actual combination of dances (on the different
floors) that actually restore the floor to the original state.

The period of that permutation can then be used to find the remainder
in the fashion described above. This was a clever design by the author.