🎄 day-17 part 2
Went out and when I came back decided to do a little hands-on experiment
with the position of particular entries.

1 does not always stay at the end of the list. So I thought all of the
other numbers would eventually move forward in the list. Why not.

Turns out: 0 always stays at the beginning of the list. Pretty cool eh?

I computed the pairs of numbers and their insert position. Filtered the
ones that have position == 0 to see them. None appeared.

Well then, fine, let's check for those that have position == 1.

Oh now some appear, cool! So the last time I do this is going to be the
final value in the 1 spot. Boom. Who'd have thunk.

The triple means (size of memory, insert position, number to insert)

    $ time ./Smrt
    (2,1,1)
    (3,1,2)
    (4,1,3)
    (5,1,4)
    (6,1,5)
    (7,1,6)
    (12,1,11)
    (13,1,12)
    (27,1,26)
    (169,1,168)
    (192,1,191)
    (262,1,261)
    (532,1,531)
    (2617,1,2616)
    (117909,1,117908)
    (182478,1,182477)
    (455679,1,455678)
    (2345316,1,2345315)
    (6693249,1,6693248)
    (9190834,1,9190833)
    (22074524,1,22074523)
    (39479737,1,39479736)
    
    real    0m0.993s
    user    0m0.885s
    sys     0m0.033s