🎄 day-23 part 2 -- primes 😮
Turns out writing an efficient virtual machine was the wrong approach.
I kept it because it does program termination.

Maybe I should backport it to day-18's code. Anyway…

Hand-translating the ASM into C and running it worked well for part 1.

Then hand-translating the version with a=1 for part 2 also did the trick
because it became clear that what the code was doing was counting the
non-primes inside the sequence [108100,108100+17..108100+17000].

The rest was elementary! 🤓